+++ src/gen-charts.ts
@@ -549,7 +549,7 @@ export function makeXmlCharts(rel: ISlideRelChart): string {
 			strXml += '	  <a:lstStyle/>'
 			strXml += '	  <a:p>'
 			strXml += '		<a:pPr rtl="0">'
-			strXml += '		  <a:defRPr sz="1197" b="0" i="0" u="none" strike="noStrike" kern="1200" baseline="0">'
+			strXml += `       <a:defRPr sz="${rel.opts.dataTableFontSize || DEF_FONT_SIZE}00" b="0" i="0" u="none" strike="noStrike" kern="1200" baseline="0">`
 			strXml += '			<a:solidFill><a:schemeClr val="tx1"><a:lumMod val="65000"/><a:lumOff val="35000"/></a:schemeClr></a:solidFill>'
 			strXml += '			<a:latin typeface="+mn-lt"/>'
 			strXml += '			<a:ea typeface="+mn-ea"/>'